chore(sdk): move experimental code to a dedicated directory
