+++ experimental/client-go/go.mod
@@ -1,8 +1,57 @@
-module github.com/wandb/wandb/experimental/client-go/gowandb
+module github.com/wandb/wandb/experimental/client-go
 
-replace (
-	// github.com/wandb/wandb => /Users/dimaduev/dev/sdk
-	github.com/wandb/wandb/experimental/client-go/gowandb => /Users/dimaduev/dev/sdk/experimental/client-go
-)
+replace github.com/wandb/wandb/core => ../../core
 
 go 1.22.2
+
+require (
+	github.com/segmentio/encoding v0.4.0
+	github.com/wandb/wandb/core v0.0.0-00010101000000-000000000000
+	google.golang.org/protobuf v1.34.0
+)
+
+require (
+	dario.cat/mergo v1.0.0 // indirect
+	github.com/Khan/genqlient v0.7.0 // indirect
+	github.com/Microsoft/go-winio v0.6.1 // indirect
+	github.com/NVIDIA/go-nvml v0.12.0-3 // indirect
+	github.com/ProtonMail/go-crypto v1.0.0 // indirect
+	github.com/cloudflare/circl v1.3.7 // indirect
+	github.com/cyphar/filepath-securejoin v0.2.4 // indirect
+	github.com/emirpasic/gods v1.18.1 // indirect
+	github.com/getsentry/sentry-go v0.27.0 // indirect
+	github.com/go-git/gcfg v1.5.1-0.20230307220236-3a3c6141e376 // indirect
+	github.com/go-git/go-billy/v5 v5.5.0 // indirect
+	github.com/go-git/go-git/v5 v5.11.0 // indirect
+	github.com/go-ole/go-ole v1.3.0 // indirect
+	github.com/golang/groupcache v0.0.0-20210331224755-41bb18bfe9da // indirect
+	github.com/hashicorp/go-cleanhttp v0.5.2 // indirect
+	github.com/hashicorp/go-retryablehttp v0.7.5 // indirect
+	github.com/jbenet/go-context v0.0.0-20150711004518-d14ea06fba99 // indirect
+	github.com/kevinburke/ssh_config v1.2.0 // indirect
+	github.com/lufia/plan9stats v0.0.0-20240226150601-1dcf7310316a // indirect
+	github.com/pjbgf/sha1cd v0.3.0 // indirect
+	github.com/power-devops/perfstat v0.0.0-20240221224432-82ca36839d55 // indirect
+	github.com/radovskyb/watcher v1.0.7 // indirect
+	github.com/segmentio/asm v1.2.0 // indirect
+	github.com/sergi/go-diff v1.3.1 // indirect
+	github.com/shirou/gopsutil/v3 v3.24.2 // indirect
+	github.com/shoenig/go-m1cpu v0.1.6 // indirect
+	github.com/skeema/knownhosts v1.2.2 // indirect
+	github.com/tklauser/go-sysconf v0.3.13 // indirect
+	github.com/tklauser/numcpus v0.7.0 // indirect
+	github.com/vektah/gqlparser/v2 v2.5.11 // indirect
+	github.com/wandb/simplejsonext v0.0.0-20240325214351-2a76dcabf635 // indirect
+	github.com/xanzy/ssh-agent v0.3.3 // indirect
+	github.com/yusufpapurcu/wmi v1.2.4 // indirect
+	golang.org/x/crypto v0.22.0 // indirect
+	golang.org/x/mod v0.17.0 // indirect
+	golang.org/x/net v0.24.0 // indirect
+	golang.org/x/sync v0.7.0 // indirect
+	golang.org/x/sys v0.19.0 // indirect
+	golang.org/x/text v0.14.0 // indirect
+	golang.org/x/time v0.5.0 // indirect
+	golang.org/x/tools v0.20.0 // indirect
+	gopkg.in/warnings.v0 v0.1.2 // indirect
+	gopkg.in/yaml.v3 v3.0.1 // indirect
+)
