+++ experimental/client-go/internal/launcher/launcher.go
@@ -9,7 +9,7 @@ import (
 	"strings"
 	"time"
 
-	"github.com/wandb/wandb/experimental/client-go/gowandb/internal/execbin"
+	"github.com/wandb/wandb/experimental/client-go/internal/execbin"
 )
 
 // readLines reads a whole file into memory
