+++ experimental/client-go/pkg/opts/runopts/runopts.go
@@ -3,7 +3,7 @@ package runopts
 
 import (
 	"github.com/wandb/wandb/core/pkg/service"
-	"github.com/wandb/wandb/experimental/client-go/gowandb/runconfig"
+	"github.com/wandb/wandb/experimental/client-go/pkg/runconfig"
 )
 
 type RunParams struct {
